Allow for firms to be launched without bootrom lockout...
... if (firmHeader->reserved2[0] & 2)

If that's the case the firm file is loaded to 0x08080000, with a max size of 0x77000 bytes, which overwrites the temporary copy of the bootroms from 0x08080000 to 0x080A0000.

Otherwise the firm file is loaded to 0x20001000, with a max size of 0x07FFF000 bytes.

The entrypoint of arm9_stage2 and the FIRM header checks have been updated accordingly.